Rename "prime" to "cofactor", increase database size and number of hashes.
We assume that the input value (a SHA-1 hash) is uniformly distributed
over the interval [0, 2^160], so the cofactors we use don't have to be
prime.  In particular, we can just use 1 as one of the cofactors.  The
others are in fact prime, but all they really have to be is relatively
prime to DB_BITS, so that they are a generator in the group of
integers mod DB_BITS.

Double the size of the database and increase the number of cofactors
to six to reduce the false-positive probability to something we can
live with in the future where this code is actually used to screen new
user passwords.  While we're at it, make DB_BITS be prime (the first
prime greater than 2^32) and ensure that actual size of the file is
big enough to accommodate this.